Revert "update return type"
This reverts commit 466dd2ed006129bde41ab7570a15004360411fa9.

+++ src/main/java/com/revature/eval/java/core/EvaluationService.java
@@ -277,7 +277,7 @@ public class EvaluationService {
 	 * @param l
 	 * @return
 	 */
-	public List<Long> calculatePrimeFactorsOf(long l) {
+	public List<Integer> calculatePrimeFactorsOf(long l) {
 		// TODO Write an implementation for this method declaration
 		return null;
 	}
